requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Web Framework
flask>=3.0.0

# RAG System Dependencies
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-ollama>=0.1.0

# Machine Learning
transformers>=4.36.0
torch>=2.2.0
sentence-transformers>=2.5.0

# Utilities
python-dotenv>=1.0.0